refactor: 컨트롤러 로직 변경
