--- src/main/java/chess/controller/Action.java
@@ -1,10 +1,0 @@
-package chess.controller;
-
-import chess.dto.ChessRequest;
-import chess.service.ChessGameService;
-import chess.view.OutputView;
-
-@FunctionalInterface
-public interface Action {
-    GameCommand execute(ChessGameService chessGameService, OutputView outputView, ChessRequest chessRequest);
-}
